public/requirements.txt
pwntools==4.14.1
scapy==2.6.1
